Moved call to API from EventViewModel to EventRepository
